+++ RELEASES.md
@@ -1,3 +1,21 @@
+# NautilusTrader 1.135.0 Beta - Release Notes
+
+Released TBD
+
+## Breaking Changes
+None
+
+## Enhancements
+- Added `CancelAllOrders` command.
+- Added bulk cancel capability to Binance integration.
+- Added bulk cancel capability to Betfair integration.
+
+## Fixes
+- Fixed handling of `cpu_freq` call in logging for ARM architecture.
+- Fixed market order fill edge case for bar data
+
+---
+
 # NautilusTrader 1.134.0 Beta - Release Notes
 
 Released on 22nd, November 2021